added a workaround for two problems that would cause a NullPointerException in Vertex.getEdges()
getEdges() is called by JsonVisitor.visit() and if the source vertex of
the edges were not set correctly when creating the graph (i.e.
manually), getSource() will be null.
getEdges() is called by JsonVisitor.parseJsonString() right before the
edges are fixed and the source vertex is set, so getSource() cannot be
anything else than null for the unfixed edges.